requirements.txt
streamlit
databricks-sdk
databricks-sql-connector
pandas
pyarrow